refactor(inference): extract Metal GDN layer encoder
Fixes #840
